requirements.txt
ipython==8.23.0
jupyter==1.0.0
keras==2.15.0
matplotlib==3.8.4
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
scikit-learn==1.4.1.post1
scipy==1.13.0
tensorflow==2.15.0